requirements.txt
streamlit
google-generativeai
numpy
pandas
mysqlclient
mysql-connector-python 
matplotlib 
prophet
